refactor(settings): extract chart_settings.dart
