+++ test/me/settings_test.dart
@@ -7,16 +7,9 @@ import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:flutter_test/flutter_test.dart";
 import "package:isar/isar.dart";
-import "package:quiver/time.dart";
 import "package:shared_preferences/shared_preferences.dart";
 
 void main() {
-  group("Duration.toMsString", () {
-    test("1000 ms", () => expect(aSecond.toMsString(), "1000 ms"));
-    test("1500 ms", () => expect((aSecond * 1.5).toMsString(), "1500 ms"));
-    test("2000 ms", () => expect((aSecond * 2).toMsString(), "2000 ms"));
-  });
-
   group("Settings", () {
     setUp(() async {
       SharedPreferences.setMockInitialValues({K.showDevTools: true});
